requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
phonenumbers==8.13.46
XlsxWriter==3.2.0
